portability: static-build: try haskell.nix (WIP)
I've just copied the flake.nix I wrote here
https://gitlab.iscpif.fr/gargantext/haskell-gargantext/merge_requests/258
and adapted it to github-action-scan.

So far I've not removed everything that can be (far from it)
and just went straight to something that could make `nix -L develop` work.
This will take the night.

If this ever works, then I'll turn to the next demi-boss:

```console
$ nix -L build
```

And then the final boss:

```console
$ nix -L build .#static-x86_64-unknown-linux-musl:github-action-scan:exe:github-action-scan-exe
```

Zzz